langchain-project/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-openai>=0.0.5
openai>=1.8.0
chromadb>=0.4.22
tiktoken>=0.5.2
langgraph>=0.0.15
unstructured>=0.11.2
pypdf>=3.17.1
python-magic>=0.4.27
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
pdfminer.six>=20221105
fastapi>=0.115.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.31.0
plotly>=5.18.0
httpx>=0.25.0
supabase>=2.3.0
watchdog>=3.0.0 